Cells data set

Data set taken from 2012_Crawleybellreg and posteriorly analyzed by 2020_Lemontebellreg. The data includes the count of infected blood cells per square millimetre on microscope slides prepared from n = 511 randomly selected individuals.

Format

A data frame with 511 rows and 5 variables:

cells: count of infected blood cells per square millimetre on microscope slides

smoker: smoking status of the subject (0: smoker; 1: non smoker)

gender: subject's gender (1: male; 0: female).

age: subject's age categorized into three levels: young (\( \le 20\)), mid (21 to 59), and old (\(\ge 60\)).

weight: body mass score categorized into three levels: normal, overweight, obese.

References

2012_Crawleybellreg

2020_Lemontebellreg